+++ examples/typescript-esm/sponsored_transactions/server_as_sponsor.ts
@@ -18,6 +18,7 @@ import {
   Network,
   NetworkToNetworkName,
   RawTransaction,
+  SimpleTransaction,
 } from "@aptos-labs/ts-sdk";
 
 const INITIAL_BALANCE = 100_000_000;
@@ -30,24 +31,24 @@ const config = new AptosConfig({ network: APTOS_NETWORK });
 const aptos = new Aptos(config);
 
 // The sponsor server gets the serialized transaction to sign as the fee payer
-const sendToTheSponsorServer = async (rawTransactionBytes: Uint8Array) => {
+const sendToTheSponsorServer = async (transactionBytes: Uint8Array) => {
   const sponsor = Account.generate();
   console.log(`Sponsor's address is: ${sponsor.accountAddress}`);
   await aptos.fundAccount({ accountAddress: sponsor.accountAddress, amount: INITIAL_BALANCE });
 
   // deserialize raw transaction
-  const deserializer = new Deserializer(rawTransactionBytes);
-  const rawTransaction = RawTransaction.deserialize(deserializer);
+  const deserializer = new Deserializer(transactionBytes);
+  const transaction = SimpleTransaction.deserialize(deserializer);
 
   // Sponsor signs
   const sponsorAuth = aptos.transaction.signAsFeePayer({
     signer: sponsor,
-    transaction: { rawTransaction, feePayerAddress: sponsor.accountAddress },
+    transaction,
   });
 
   const sponsorAuthBytes = sponsorAuth.bcsToBytes();
 
-  return { sponsorAuthBytes, feepayerAddress: sponsor.accountAddress };
+  return sponsorAuthBytes;
 };
 
 const example = async () => {
@@ -81,26 +82,15 @@ const example = async () => {
   // Alice signs
   const senderAuth = aptos.transaction.sign({ signer: alice, transaction });
 
-  // Serialize data
-  /**
-   * the generated `transaction` is of a simple type and can't be
-   * bcs serialized. Therefore we serialize the `rawTransaction` and
-   * send it to the server.
-   */
-  const rawTransactionBytes = transaction.rawTransaction.bcsToBytes();
-
-  // Send the transaction to the sponsor server to sign
-  const { sponsorAuthBytes, feepayerAddress } = await sendToTheSponsorServer(rawTransactionBytes);
+  // Send the serialized transaction to the sponsor server to sign
+  const sponsorAuthBytes = await sendToTheSponsorServer(transaction.bcsToBytes());
 
   // deserialize fee payer authenticator
   const deserializer3 = new Deserializer(sponsorAuthBytes);
   const feePayerAuthenticator = AccountAuthenticator.deserialize(deserializer3);
 
   const response = await aptos.transaction.submit.simple({
-    transaction: {
-      rawTransaction: transaction.rawTransaction,
-      feePayerAddress: feepayerAddress,
-    },
+    transaction,
     senderAuthenticator: senderAuth,
     feePayerAuthenticator,
   });