+++ src/internal/digitalAsset.ts
@@ -11,7 +11,7 @@
 import { AptosConfig } from "../api/aptosConfig";
 import { Bool, MoveString, MoveVector, U64 } from "../bcs";
 import { Account, AccountAddress, AccountAddressInput } from "../core";
-import { EntryFunctionABI, InputGenerateTransactionOptions, SimpleTransaction } from "../transactions/types";
+import { EntryFunctionABI, InputGenerateTransactionOptions } from "../transactions/types";
 import {
   AnyNumber,
   GetCollectionDataResponse,
@@ -53,6 +53,7 @@ import {
   TypeTagU64,
   TypeTagVector,
 } from "../transactions";
+import { SimpleTransaction } from "../transactions/instances/simpleTransaction";
 
 // A property type map for the user input and what Move expects
 const PropertyTypeMap = {