+++ src/transactions/types.ts
@@ -18,6 +18,8 @@ import {
 import { AnyNumber, HexInput, MoveFunctionGenericTypeParam, MoveFunctionId } from "../types";
 import { TypeTag } from "./typeTag";
 import { AccountAuthenticator } from "./authenticator/account";
+import { SimpleTransaction } from "./instances/simpleTransaction";
+import { MultiAgentTransaction } from "./instances/multiAgentTransaction";
 
 /**
  * Entry function arguments to be used when building a raw transaction using remote ABI
@@ -223,29 +225,6 @@ export interface InputGenerateMultiAgentRawTransactionArgs {
 export type InputGenerateRawTransactionArgs =
   | InputGenerateSingleSignerRawTransactionArgs
   | InputGenerateMultiAgentRawTransactionArgs;
-
-/**
- * Interface that holds the return data when generating a single signer transaction
- *
- * @param rawTransaction a bcs serialized raw transaction
- */
-export interface SimpleTransaction {
-  rawTransaction: RawTransaction;
-  feePayerAddress?: AccountAddress;
-  secondarySignerAddresses?: undefined;
-}
-
-/**
- * Interface that holds the return data when generating a multi-agent transaction.
- *
- * @param rawTransaction a bcs serialized raw transaction
- * @param secondarySignerAddresses secondary signer addresses for multi-agent transaction
- */
-export interface MultiAgentTransaction {
-  rawTransaction: RawTransaction;
-  secondarySignerAddresses: AccountAddress[];
-  feePayerAddress?: AccountAddress;
-}
 
 /**
  * Unified type that holds all the return interfaces when generating different transaction types